# packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vulndesk API",
    "version": "0.1.0",
    "description": "CVE / security-advisory authoring API. Building in public — v1: @vulndesk/core validation + document CRUD."
  },
  "components": {
    "schemas": {
      "ValidationResult": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean"
          },
          "errorCount": {
            "type": "integer"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "valid",
          "errorCount",
          "errors"
        ]
      },
      "CveRecord": {
        "type": "object",
        "additionalProperties": {},
        "description": "A CVE Record Format 5.x record (JSON object)."
      },
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "section": {
            "type": "string"
          },
          "docId": {
            "type": "string"
          },
          "cveId": {
            "type": [
              "string",
              "null"
            ]
          },
          "state": {
            "type": [
              "string",
              "null"
            ]
          },
          "author": {
            "type": [
              "string",
              "null"
            ]
          },
          "body": {
            "$ref": "#/components/schemas/CveRecord"
          },
          "slug": {
            "type": [
              "string",
              "null"
            ]
          },
          "fullSlug": {
            "type": [
              "string",
              "null"
            ]
          },
          "version": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "section",
          "docId",
          "cveId",
          "state",
          "author",
          "body",
          "slug",
          "fullSlug",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "error"
        ]
      },
      "CreateDocumentInput": {
        "type": "object",
        "properties": {
          "section": {
            "type": "string",
            "minLength": 1,
            "example": "cve5"
          },
          "body": {
            "$ref": "#/components/schemas/CveRecord"
          },
          "docId": {
            "type": "string",
            "description": "Lookup id; derived from the body (e.g. cveMetadata.cveId) when omitted."
          },
          "author": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "fullSlug": {
            "type": "string"
          }
        },
        "required": [
          "section",
          "body"
        ]
      },
      "DocumentList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Document"
            }
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "UpdateDocumentInput": {
        "type": "object",
        "properties": {
          "body": {
            "$ref": "#/components/schemas/CveRecord"
          },
          "state": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "fullSlug": {
            "type": "string"
          }
        }
      }
    },
    "parameters": {}
  },
  "paths": {
    "/validate": {
      "post": {
        "tags": [
          "CVE"
        ],
        "summary": "Validate a CVE Record Format 5.x record",
        "description": "Validates the record against the official CVE5 JSON Schema via @vulndesk/core.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CveRecord"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            }
          }
        }
      }
    },
    "/cve5/schema": {
      "get": {
        "tags": [
          "CVE"
        ],
        "summary": "The canonical CVE Record Format 5.x JSON Schema",
        "responses": {
          "200": {
            "description": "CVE5 JSON Schema",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
    "/documents": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Create a document",
        "description": "Stores a CVE/advisory record (drafts allowed — use /validate to check conformance). docId/cveId/state are promoted from the body for indexing.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate (section, docId)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "List documents",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "section",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "state",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "cveId",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of documents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentList"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get a document by id",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "00000000-0000-0000-0000-000000000000"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Documents"
        ],
        "summary": "Update a document",
        "description": "Patches the supplied fields and bumps version. A new body re-derives cveId/state.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "00000000-0000-0000-0000-000000000000"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Documents"
        ],
        "summary": "Delete a document",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "00000000-0000-0000-0000-000000000000"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "webhooks": {}
}